--- src/Codeworx.Rest.Formatters.Newtonsoft/PublicAPI.Shipped.txt
@@ -1,1 +1,0 @@
-﻿
