+++ test/Codeworx.Rest.IntegrationTests/Codeworx.Rest.UnitTests.csproj
@@ -6,7 +6,7 @@
 		<TreatWarningsAsErrors>false</TreatWarningsAsErrors>
 	</PropertyGroup>
 	<ItemGroup>
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.5.0" />
+		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.1" />
 		<PackageReference Include="NSwag.AspNetCore" Version="13.18.2" />
 		<PackageReference Include="xunit" Version="2.4.2" />
 		<PackageReference Include="xunit.runner.visualstudio" Version="2.4.5">
@@ -25,17 +25,10 @@
 		<PackageReference Include="Microsoft.AspNetCore.TestHost" Version="7.0.0" />
 		<PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="7.0.0" />
 	</ItemGroup>
-		<ItemGroup>
-		  <Compile Include="..\..\src\Codeworx.Rest.Client\Internal\CodeworxRestClientInternalServiceCollectionExtensions.cs" Link="Internal\CodeworxRestClientInternalServiceCollectionExtensions.cs" />
-		</ItemGroup>
 	<ItemGroup>
 		<ProjectReference Include="..\..\src\Codeworx.Rest.AspNetCore\Codeworx.Rest.AspNetCore.csproj" />
 		<ProjectReference Include="..\..\src\Codeworx.Rest.Client\Codeworx.Rest.Client.csproj" />
-		<ProjectReference Include="..\..\src\Codeworx.Rest.Formatters.Newtonsoft\Codeworx.Rest.Formatters.Newtonsoft.csproj" />
 		<ProjectReference Include="..\..\src\Codeworx.Rest.Formatters.Protobuf\Codeworx.Rest.Formatters.Protobuf.csproj" />
 		<ProjectReference Include="..\Codeworx.Rest.UnitTests.Api.Contract\Codeworx.Rest.UnitTests.Api.Contract.csproj" />
-	</ItemGroup>
-	<ItemGroup>
-	  <Folder Include="Internal\" />
 	</ItemGroup>
 </Project>
