+++ test/Codeworx.Rest.IntegrationTests/FormatterSelection.cs
@@ -1,9 +1,12 @@
-﻿namespace Codeworx.Rest.UnitTests
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Codeworx.Rest.UnitTests
 {
     public enum FormatterSelection
     {
         Json = 0x00,
-        Protobuf = 0x01,
-        NewtonsoftJson = 0x2
+        Protobuf = 0x01
     }
 }
