+++ test/Codeworx.Rest.IntegrationTests/SerializeResultTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Codeworx.Rest.UnitTests.Api.Contract;
 using Codeworx.Rest.UnitTests.Data;
+using Codeworx.Rest.UnitTests.Generated;
 using Xunit;
 
 namespace Codeworx.Rest.UnitTests
