+++ test/Codeworx.Rest.IntegrationTests/TestServerTestsBase.cs
@@ -93,10 +93,6 @@ namespace Codeworx.Rest.UnitTests
             {
                 builder = builder.AddProtobufFormatter(makeDefault: true);
             }
-            else if (selection == FormatterSelection.NewtonsoftJson)
-            {
-                builder = builder.AddNewtonsoftJsonFormatter();
-            }
 
             builderAction?.Invoke(builder);
