+++ test/Codeworx.Rest.UnitTests.Api.Contract/Model/Item.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Runtime.Serialization;
 
 namespace Codeworx.Rest.UnitTests.Model
