+++ tools/build.ps1
@@ -4,7 +4,6 @@ Import-Module -Name "./Build-Versioning.psm1"
 $projects = "..\src\Codeworx.Rest.Primitives\Codeworx.Rest.Primitives.csproj",
             "..\src\Codeworx.Rest.Client\Codeworx.Rest.Client.csproj", 
             "..\src\Codeworx.Rest.Formatters.Protobuf\Codeworx.Rest.Formatters.Protobuf.csproj", 
-            "..\src\Codeworx.Rest.Formatters.Newtonsoft\Codeworx.Rest.Formatters.Newtonsoft.csproj", 
             "..\src\Codeworx.Rest.AspNetCore\Codeworx.Rest.AspNetCore.csproj", 
             "..\src\Codeworx.Rest.Tool.Cli\Codeworx.Rest.Tool.Cli.csproj"
