+++ version.json
@@ -1,5 +1,5 @@
 {
-    "major": 2,
+    "major": 1,
     "minor": 0,
     "buildNumberPrefix": 0
 }
